+++ crates/atlassian-swc-compiled-css/tests/fixtures/css-jsx-wrong-pragma/out.js
@@ -1,5 +1,0 @@
-<div
-  css={{
-    color: 'red',
-  }}
-/>;