+++ packages/packagers/js/src/utils.ts
@@ -1,5 +1,5 @@
 import type {BundleGraph, Dependency, NamedBundle} from '@atlaspack/types';
-import type SourceMap from '@parcel/source-map';
+import type SourceMap from '@atlaspack/source-map';
 import nullthrows from 'nullthrows';
 
 // This replaces __parcel__require__ references left by the transformer with