+++ packages/transformers/js/core/src/lib.rs
@@ -105,7 +105,7 @@ use swc_core::ecma::visit::VisitMutWith;
 use swc_core::ecma::visit::VisitWith;
 use swc_core::ecma::visit::visit_mut_pass;
 use sync_dynamic_import::SyncDynamicImport;
-use sync_dynamic_import::SyncDynamicImportConfig;
+pub use sync_dynamic_import::SyncDynamicImportConfig;
 use typeof_replacer::*;
 use unused_bindings_remover::UnusedBindingsRemover;
 use utils::CodeHighlight;
@@ -227,7 +227,7 @@ fn targets_to_versions(targets: &Option<HashMap<String, String>>) -> Option<Vers
 }
 
 pub fn transform(
-  config: Config,
+  config: &Config,
   call_macro: Option<MacroCallback>,
 ) -> Result<TransformResult, io::Error> {
   let mut result = TransformResult::default();
@@ -240,7 +240,7 @@ pub fn transform(
     config.project_root.as_str(),
     config.filename.as_str(),
     &source_map,
-    &config,
+    config,
   );
 
   match module {
@@ -461,6 +461,13 @@ pub fn transform(
                     config.source_type != SourceType::Script
                   ),
                   Optional::new(
+                    visit_mut_pass(
+                      SyncDynamicImport::new(Path::new(&config.filename),
+                        unresolved_mark,
+                        &config.sync_dynamic_import_config,
+                      )),
+                      config.sync_dynamic_import_config.is_some()),
+                  Optional::new(
                     visit_mut_pass(GlobalAliaser::with_config(unresolved_mark, &config.global_aliasing_config)),
                     config.global_aliasing_config.is_some()
                   ),
@@ -586,7 +593,7 @@ pub fn transform(
                       &mut result.dependencies,
                       ignore_mark,
                       unresolved_mark,
-                      &config,
+                      config,
                       &mut diagnostics,
                       &mut result.conditions,
                     ),
@@ -874,12 +881,12 @@ mod tests {
   fn test_logs_when_flag_is_on_and_file_is_empty() {
     let config: Config = make_test_swc_config(r#""#);
     unsafe { env::set_var("ATLASPACK_SHOULD_LOOK_FOR_EMPTY_FILES", "true") };
-    let _result = transform(config, None);
+    let _result = transform(&config, None);
     assert!(logs_contain("You are attempting to import"));
 
     unsafe { env::set_var("ATLASPACK_SHOULD_LOOK_FOR_EMPTY_FILES", "false") };
     let config: Config = make_test_swc_config(r#""#);
-    let _result = transform(config, None);
+    let _result = transform(&config, None);
     logs_assert(|lines: &[&str]| {
       let count = lines
         .iter()