+++ packages/transformers/tokens/src/TokensTransformer.ts
@@ -3,7 +3,7 @@ import {getFeatureFlag} from '@atlaspack/feature-flags';
 import {Transformer} from '@atlaspack/plugin';
 import {applyTokensPlugin, TokensPluginResult} from '@atlaspack/rust';
 import {validateSchema} from '@atlaspack/utils';
-import SourceMap from '@parcel/source-map';
+import SourceMap from '@atlaspack/source-map';
 import path from 'path';
 
 type AtlaskitTokensConfigPartial = {
@@ -46,9 +46,7 @@ export default new Transformer({
         CONFIG_SCHEMA,
         {
           data: conf.contents,
-          source: getFeatureFlag('schemaValidationDeferSourceLoading')
-            ? () => options.inputFS.readFileSync(conf.filePath, 'utf8')
-            : await options.inputFS.readFile(conf.filePath, 'utf8'),
+          source: () => options.inputFS.readFileSync(conf.filePath, 'utf8'),
           filePath: conf.filePath,
           prependKey: `/${encodeJSONKeyComponent('@atlaspack/transformer-tokens')}`,
         },
@@ -112,7 +110,6 @@ export default new Transformer({
       let map = new SourceMap(options.projectRoot);
       map.addVLQMap(JSON.parse(result.map));
       if (originalMap) {
-        // @ts-expect-error TS2345 - the types are wrong, `extends` accepts a `SourceMap` or a `Buffer`
         map.extends(originalMap);
       }
       asset.setMap(map);