+++ packages/compressors/raw/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaspack/compressor-raw
 
+## 2.13.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/plugin@2.14.40
+
+## 2.13.41
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/plugin@2.14.39
+
 ## 2.13.40
 
 ### Patch Changes