+++ packages/resolvers/default/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaspack/resolver-default
 
+## 2.14.40
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/node-resolver-core@3.7.11
+  - @atlaspack/plugin@2.14.40
+
+## 2.14.39
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/node-resolver-core@3.7.10
+  - @atlaspack/plugin@2.14.39
+
 ## 2.14.38
 
 ### Patch Changes